added file "AnimalClassify.h" where I defined an enum for animal classes.
implemented the method "addAnimalsToZoo" using the enum, and by defining a method which matches an area to a specific habitat(animal class).

left "addAnimalsToZoo" with an erro because we need to figure out what to do with the method "getAllAreas" which return a const - trouble when trying to add an animal to that area!